Refactor consumer/generator/utils
This commit splits the generator into three parts:
 * ASTConsumer: consumes the AST, using the generator for all MLIR
   purposes.
 * MLIRGenerator: responsible for generating MLIR code and keeping the
   context between AST-generated and compiler-generated IR.
 * Utils: a simple collection of AST/MLIR helpers used on both.

The purpose of this split is to be able to use the generator from passes
as well as the AST consumer, so that we don't duplicate code.
