+++ testsuite/verona-mlir/empty-class.out/stdout.txt
@@ -1,2 +1,2 @@
-module @__  {
+module  {
 }
