+++ testsuite/verona-mlir/function.out/stdout.txt
@@ -1,12 +1,12 @@
-module @__  {
-  func @"____$module-0__apply"() {
+module  {
+  func @"$module-0__apply"() {
     return
   }
-  func @"____$module-0__retOnly"() -> i64 {
+  func @"$module-0__retOnly"() -> i64 {
     %c42_i64 = constant 42 : i64
     return %c42_i64 : i64
   }
-  func @"____$module-0__retOneArg"(%arg0: i64) -> i64 {
+  func @"$module-0__retOneArg"(%arg0: i64) -> i64 {
     %c1_i32 = constant 1 : i32
     %0 = llvm.alloca %c1_i32 x i64 : (i32) -> !llvm.ptr<i64>
     %c0_i32 = constant 0 : i32
@@ -17,7 +17,7 @@ module @__  {
     %3 = llvm.load %2 : !llvm.ptr<i64>
     return %3 : i64
   }
-  func @"____$module-0__retFirstArg"(%arg0: i64, %arg1: i64) -> i64 {
+  func @"$module-0__retFirstArg"(%arg0: i64, %arg1: i64) -> i64 {
     %c1_i32 = constant 1 : i32
     %0 = llvm.alloca %c1_i32 x i64 : (i32) -> !llvm.ptr<i64>
     %c0_i32 = constant 0 : i32
@@ -33,7 +33,7 @@ module @__  {
     %5 = llvm.load %4 : !llvm.ptr<i64>
     return %5 : i64
   }
-  func @"____$module-0__retSecondArg"(%arg0: i64, %arg1: i64) -> i64 {
+  func @"$module-0__retSecondArg"(%arg0: i64, %arg1: i64) -> i64 {
     %c1_i32 = constant 1 : i32
     %0 = llvm.alloca %c1_i32 x i64 : (i32) -> !llvm.ptr<i64>
     %c0_i32 = constant 0 : i32
@@ -49,7 +49,7 @@ module @__  {
     %5 = llvm.load %4 : !llvm.ptr<i64>
     return %5 : i64
   }
-  func @"____$module-0__retSecondArgDiff"(%arg0: i64, %arg1: i32) -> i32 {
+  func @"$module-0__retSecondArgDiff"(%arg0: i64, %arg1: i32) -> i32 {
     %c1_i32 = constant 1 : i32
     %0 = llvm.alloca %c1_i32 x i64 : (i32) -> !llvm.ptr<i64>
     %c0_i32 = constant 0 : i32
