+++ testsuite/verona-mlir/main.out/stdout.txt
@@ -1,5 +1,5 @@
-module @__  {
-  func @"____$module-0__Math__getTruth"() -> i64 {
+module  {
+  func @"$module-0__Math__getTruth"() -> i64 {
     %c42_i64 = constant 42 : i64
     %c1_i32 = constant 1 : i32
     %0 = llvm.alloca %c1_i32 x i64 : (i32) -> !llvm.ptr<i64>
@@ -14,7 +14,7 @@ module @__  {
     %5 = llvm.load %4 : !llvm.ptr<i64>
     return %5 : i64
   }
-  func @"____$module-0__Math__getRandom"() -> i64 {
+  func @"$module-0__Math__getRandom"() -> i64 {
     %c1_i64 = constant 1 : i64
     %c1_i32 = constant 1 : i32
     %0 = llvm.alloca %c1_i32 x i64 : (i32) -> !llvm.ptr<i64>
@@ -29,10 +29,10 @@ module @__  {
     %5 = llvm.load %4 : !llvm.ptr<i64>
     return %5 : i64
   }
-  func @"____$module-0__bar"() -> i64 {
+  func @"$module-0__bar"() -> i64 {
     %c1_i32 = constant 1 : i32
     %0 = llvm.alloca %c1_i32 x i64 : (i32) -> !llvm.ptr<i64>
-    %1 = call @"____$module-0__Math__getRandom"() : () -> i64
+    %1 = call @"$module-0__Math__getRandom"() : () -> i64
     %c0_i32 = constant 0 : i32
     %2 = llvm.getelementptr %0[%c0_i32] : (!llvm.ptr<i64>, i32) -> !llvm.ptr<i64>
     %3 = llvm.load %2 : !llvm.ptr<i64>
@@ -44,7 +44,7 @@ module @__  {
     %6 = llvm.load %5 : !llvm.ptr<i64>
     return %6 : i64
   }
-  func @"____$module-0__foo"(%arg0: i64) -> i64 {
+  func @"$module-0__foo"(%arg0: i64) -> i64 {
     %c1_i32 = constant 1 : i32
     %0 = llvm.alloca %c1_i32 x i64 : (i32) -> !llvm.ptr<i64>
     %c0_i32 = constant 0 : i32
@@ -52,7 +52,7 @@ module @__  {
     llvm.store %arg0, %1 : !llvm.ptr<i64>
     %c1_i32_0 = constant 1 : i32
     %2 = llvm.alloca %c1_i32_0 x i64 : (i32) -> !llvm.ptr<i64>
-    %3 = call @"____$module-0__bar"() : () -> i64
+    %3 = call @"$module-0__bar"() : () -> i64
     %c0_i32_1 = constant 0 : i32
     %4 = llvm.getelementptr %2[%c0_i32_1] : (!llvm.ptr<i64>, i32) -> !llvm.ptr<i64>
     %5 = llvm.load %4 : !llvm.ptr<i64>
@@ -82,7 +82,7 @@ module @__  {
   func @main() -> i64 {
     %c1_i32 = constant 1 : i32
     %0 = llvm.alloca %c1_i32 x !llvm.struct<"Math", (i64, i64)> : (i32) -> !llvm.ptr<struct<"Math", (i64, i64)>>
-    %1 = call @"____$module-0__Math__getTruth"() : () -> i64
+    %1 = call @"$module-0__Math__getTruth"() : () -> i64
     %c1_i32_0 = constant 1 : i32
     %2 = llvm.alloca %c1_i32_0 x i64 : (i32) -> !llvm.ptr<i64>
     %c0_i32 = constant 0 : i32
@@ -125,7 +125,7 @@ module @__  {
     %c0_i32_13 = constant 0 : i32
     %22 = llvm.getelementptr %18[%c0_i32_13] : (!llvm.ptr<i64>, i32) -> !llvm.ptr<i64>
     %23 = llvm.load %22 : !llvm.ptr<i64>
-    %24 = call @"____$module-0__foo"(%23) : (i64) -> i64
+    %24 = call @"$module-0__foo"(%23) : (i64) -> i64
     %c0_i32_14 = constant 0 : i32
     %25 = llvm.getelementptr %17[%c0_i32_14] : (!llvm.ptr<i64>, i32) -> !llvm.ptr<i64>
     %26 = llvm.load %25 : !llvm.ptr<i64>
